Update configuration scope for Lombok
Gradle 4.7 changed the way annotation processing works. I'm not sure
exactly how it changed, but the effect is that I needed to change
the dependency configurations that Lombok uses. It now needs to be applied in
the annotationProcessor block and compileOnly blocks. It's only deprecated
now, but it will be removed in Gradle 5. Might as well get it done now.